pyproject.toml
[tool.poetry]
name = "storage-management"
version = "0.1.0"
description = ""
authors = ["Andrew LeCody <andrew@beyondweb.net>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
humanize = "^4.0.0"
requests = "^2.27.1"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
